introduce-multimodal-processor-move: Hand EPD dispatch over to MultimodalProcessor (with privacy flip)
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod
_should_dispatch_to_encoder + _handle_epd_disaggregation_encode_request
from TokenizerManager; paste into MultimodalProcessor (drop @staticmethod,
replace ``self: "MultimodalProcessor"`` → plain ``self``). Privacy flip
per design:
  _should_dispatch_to_encoder              -> should_dispatch_to_encoder
  _handle_epd_disaggregation_encode_request -> maybe_dispatch_to_encoder

Caller prefix replacement:
  TokenizerManager._handle_epd_disaggregation_encode_request(self.multimodal_processor, obj)
    -> self.multimodal_processor.maybe_dispatch_to_encoder(obj)
  TokenizerManager._should_dispatch_to_encoder(self, obj) (intra-class call)
    -> self.should_dispatch_to_encoder(obj)

Refactor chain ID: introduce-multimodal-processor-move
